ETH/tg_dashboard_project/requirements.txt
python-telegram-bot
psycopg2
pandas
schedule
